fixed bug in generate receipt view

+++ receipt_generator_api/models.py
@@ -69,6 +69,6 @@ class Receipt(models.Model):
     name = models.CharField(max_length=255, null=True)
     address = models.CharField(max_length=255, null =True)
     mobile_number = PhoneNumberField(null=True)
-    total_amount_payable=models.DecimalField( max_digits=50, decimal_places=2)
+    total_amount_payable=models.FloatField()
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
